Refactor: colliderBoxes.js 파일 분리
