+++ src/utils/colliderBoxes.js
@@ -1,0 +1,24 @@
+const colliderBoxes = [
+  { id: "collider1", size: [77, 20, 10], position: [-130, -27, -36] },
+  { id: "collider2", size: [130, 20, 3], position: [-130, -27, 20] },
+  {
+    id: "collider3",
+    size: [5, 20, 65],
+    position: [-174, -27, -10],
+    rotation: [0, Math.PI / 1.06, 0],
+  },
+  {
+    id: "colider4",
+    size: [3, 20, 56],
+    position: [-160, -27, 49],
+    rotation: [0, Math.PI / -1.28, 0],
+  },
+  {
+    id: "colider5",
+    size: [3, 20, 49],
+    position: [-117, -27, 70],
+    rotation: [0, Math.PI / -2, 0],
+  },
+];
+
+export default colliderBoxes;
